fix bug update cart and delete cart
